+++ src/app.module.ts
@@ -4,7 +4,7 @@ import { APP_FILTER, APP_INTERCEPTOR } from '@nestjs/core';
 import { NestResponseInterceptor } from './core/http/nestResponse.interceptor';
 import { HttpExceptionFilter } from './common/filters/httpException.filter';
 
-import { PrismaService } from './prisma/prisma.service';
+import { PrismaService } from '../prisma/prisma.service';
 import { UsersModule } from './models/users/users.module';
 import { SendMailModule } from './mail/send-mail.module';
 import { AuthModule } from './auth/auth.module';
