+++ src/auth/auth.module.ts
@@ -5,7 +5,7 @@ import { AuthController } from './auth.controller';
 import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { LocalAuthGuard } from '../guards/local.guard';
-import { PrismaService } from '../prisma/prisma.service';
+import { PrismaService } from '../../prisma/prisma.service';
 import { LocalStrategy } from '../guards/local.strategy';
 import { JwtStrategy } from '../guards/jwt.strategy';
 import { EncryptData } from '../utils/encrypt-data';
