+++ src/models/users/repository/user.repository.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { PrismaService } from '../../../prisma/prisma.service';
+import { PrismaService } from '../../../../prisma/prisma.service';
 import { CreateUserDto } from '../dto/create-user.dto';
 import { SetRoleDto } from '../dto/set-role-dto';
 import { UpdateUserDto } from '../dto/update-user.dto';
